Applied patch from Reuben Thomas:
Revert patch from Olivier Andrieu (1.36) to support long strings and
comments, as it is problematic: [[ and ]] simply cause string
colouring mode to be flipped into and out of, which means that for
example "a[b[i]]" causes string colouring from the second "]" onwards.
This is not easy to fix except by inserting extra spaces at these
points. I've asked the contributor to supply a version of the patch
which can be turned on by a config variable as a workaround solution.

Fix lua-beginning-of-proc when point is over "function (". Thanks to
Law Chi Ying for the report.