Indentation fixes and tidy up.
- The standalone do/end block was not handled correctly in
lua-find-matching-token-word. Splitting the matching token regexps for
search direction allowed a fix.
- Repeated car/cdr compination to access token info fields was not helping
maintenance. Added a few accessor functions.
- When trying to find a matching token, the search was allowed to go on
until it reached the beginning of buffer. This caused a recursion stack
overflow. Now the search terminates when it reaches a point where it is
meaningless to continue further.
- When search direction is not supplied, middle-or-open type tokens search
forward by default
- Specifying serch direction for same-line block closer handling in
lua-make-indentation-info-pair.