Fix multiline literal highlighting for long strings (#20)
lua-put-char-property & lua-unmark-multiline-region updated character properties (and
ther syntactic meaning) silently, i.e. none of the buffer modification hooks were run.
Among the disabled hooks there was the one that notified syntax-ppss of the changes
and that initiated an update of syntax-ppss's internal syntax cache. Disabling that
caused discrepancy and the function lua-comment-or-string-p started returning wrong
values, which resulted in an almost random highlighting of multiline literals.