Rewrite indentation engine
The main reason for this change is to change indentation from

  foobar('arg1', function ()
           print('foobar')
                 end)

to

  foobar('arg1', function ()
      print('foobar')
  end)

The indentation engine was very convoluted, hence the major changes.

Algorithm: indentation is computed from current line and previous line.
Current line:
-1 on every unmatched closer if first token is a closer
-1 if first token is 'middle'
Previous line:
+1 on every unmatched opener
+1 if first token is a 'middle'
-1 on every unmatched closer if first token is not a closer (if
 first token is a 'middle', then first unmatched closer is actually closing
 the 'middle').
To this we add
+ previous line indentation
+1 if previous line is not a continuation and current-line is
-1 if previous line is a continuation and current-line is not

The new indentation algorithm is as follows:
Current line indentation is:
+ indentation changes induced by unmatched closers on current line
+ indentation changes induced by previous line
+ previous line indentation
- indent-level if previous line was a continuation
- indent-level

Implementation changes:
* Keyword are only distinguished by one of the classes 'open', 'close', and 'middle'.
* Openers yield a simple indent, no alignment.
* Closers at the beginning of the line do not yield any indent for following line.
* Continuing lines are not indented together, only previous line is used for indentation.
* There is no more exceptions for 'else', 'do', and so on.
* The engine is split into 3 functions only.
* The indentation stack is stored into a single integer, there is no more list construct.

The engine should never scans back more than the previous line.
This should result in a sensible performance boost.

For now this is not the case since the function 'lua-find-matching-token-in-line'
goes beyond the boundaries of the line (quick and dirty implementation). Fixing
this should make the engine much faster.

Tests: The file test/indentation-test.el has been updated accordingly. I have
added a test for the new capabilities of the engine (among them the above
example).

Other changes include:
* Remove unused left-shifter detection code
* lua-block-token-alist: Sanitize regexp
* lua-block-token-alist: Clarify doc
* Completed some missing indentation tests.