Use cl-lib instead of cl
Emacs cl makes the latter obsolete.

Since we stopped using `lexical-let*' in the previous commit all
that is left to do is to use `cl-assert' instead of `assert'.

To increase compatibility use `nth' instead of `caddr' et al., which
depending on Emacs version and whether `cl' or `cl-lib' is loaded do
or do not have to be spelled with the `cl-' prefix.